move metadata to meta
